Use pool for Resultset, not for Result
